Simplify data format, remove hex data from flow item.
The old format is terribly inefficient, and moving it to gridFS is
slow and clunky, even if you were to manually handle appends.

Data is capped at 15 MB, to stay well under the document limit
of 16MB. Any data beyond that is discarded, but the start
of the session will remain searchable.